[treewide] lora: use int16_t for RSSI value
The RSSI values reported by LoRa transceiver can be less than -127.
Therefore, `int8_t` is not enough. This commit defines the RSSI of
`netdev_lora_rx_info` as `int16_t` and adapt the drivers accordingly
(sx126x, sx127x).

(cherry picked from commit 9955a35c639ca9441ad42d9b2838205bc30ed587)

+++ drivers/include/net/netdev/lora.h
@@ -32,7 +32,7 @@ extern "C" {
  * @brief   Received LoRa packet status information
  */
 typedef struct {
-    uint8_t rssi;           /**< RSSI of a received packet */
+    int16_t rssi;           /**< RSSI of a received packet */
     int8_t snr;             /**< S/N ratio */
 } netdev_lora_rx_info_t;
